refactor(config) move config into cmd
It won't really exist separately once command flags are incorporated,
just getting that sorted now before that change.
